small modif installer

+++ installer/TVOQE_Perso/remove_ivy.sh
@@ -5,12 +5,15 @@ done < <(find $HOME/TVOQE_Perso/QUIC-Ivy/doc/examples/quic -type f -name \*.ivy 
 
 echo $array
 
+SUB='test'
 for j in "${array[@]}"; do : 
     # Set space as the delimiter
     IFS=' '
     #Read the split words into an array based on space delimiter
     read -a strarr <<< "$j"
     printf "Files => /usr/local/lib/python2.7/dist-packages/ivy/include/1.7/$strarr  \n" 
-    sudo rm /usr/local/lib/python2.7/dist-packages/ivy/include/1.7/${strarr[0]}
+    if [[ ! "${strarr[0]}" == *"$SUB"* ]]; then
+    	sudo rm /usr/local/lib/python2.7/dist-packages/ivy/include/1.7/${strarr[0]}
+    fi
 done
